+++ enterprise/coderd/dbauthz/accesscontrol.go
@@ -1,0 +1,30 @@
+package dbauthz
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+	"golang.org/x/xerrors"
+
+	"github.com/coder/coder/v2/coderd/database"
+	agpldbz "github.com/coder/coder/v2/coderd/database/dbauthz"
+)
+
+type EnterpriseTemplateAccessControlStore struct{}
+
+func (EnterpriseTemplateAccessControlStore) GetTemplateAccessControl(t database.Template) (agpldbz.TemplateAccessControl, error) {
+	return agpldbz.TemplateAccessControl{
+		RequireActiveVersion: t.RequireActiveVersion,
+	}, nil
+}
+
+func (EnterpriseTemplateAccessControlStore) SetTemplateAccessControl(ctx context.Context, store database.Store, id uuid.UUID, opts agpldbz.TemplateAccessControl) error {
+	err := store.UpdateTemplateAccessControlByID(ctx, database.UpdateTemplateAccessControlByIDParams{
+		ID:                   id,
+		RequireActiveVersion: opts.RequireActiveVersion,
+	})
+	if err != nil {
+		return xerrors.Errorf("update template access control: %w", err)
+	}
+	return nil
+}
